+++ database/migrations/2020_03_16_103715_create_tweets_table.php
@@ -15,7 +15,7 @@ class CreateTweetsTable extends Migration
     {
         Schema::create('tweets', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->text('body')->nullable();
             $table->timestamps();
