+++ database/migrations/2020_04_27_203725_add_parent_id_to_tweets_table.php
@@ -14,7 +14,7 @@ class AddParentIdToTweetsTable extends Migration
     public function up()
     {
         Schema::table('tweets', function (Blueprint $table) {
-            $table->integer('parent_id')->unsigned()->index()->nullable();
+            $table->bigInteger('parent_id')->unsigned()->index()->nullable();
             $table->foreign('parent_id')->references('id')->on('tweets')->onDelete('cascade');
         });
     }
